Move hydrate into utils mod
