+++ packages/sycamore/src/utils/mod.rs
@@ -4,6 +4,8 @@
 //! This API is currently unstable and can have breaking changed without a semver release.
 //! This might be stabilized in the future but it is use-at-your-own-risk for now.
 
+#[cfg(feature = "hydrate")]
+pub mod hydrate;
 pub mod render;
 
 use std::cell::RefCell;
